requirements.txt
# Package dependency
# ------------------------------------------------------------------------------
numpy
scipy
pandas
scikit-learn
